factor out all extension id enums
