+++ src/commands/showInstalledVersions.ts
@@ -1,8 +1,8 @@
 import os from 'os';
 import { env, extensions, version, window } from 'vscode';
-import { failed } from '../types/errorable';
-import { GitOpsExtensionConstants } from '../extension';
 import { getAzureVersion, getFluxVersion, getGitVersion, getKubectlVersion } from '../install';
+import { failed } from '../types/errorable';
+import { GitOpsExtensionConstants } from '../types/extensionIds';
 
 /**
  * Show all installed cli versions.
