--- src/types/treeViewIds.ts
@@ -1,10 +1,0 @@
-/**
- * GitOps view ids.
- */
-export const enum TreeViews {
-	ClustersView = 'gitops.views.clusters',
-	SourcesView = 'gitops.views.sources',
-	WorkloadsView = 'gitops.views.workloads',
-	TemplatesView = 'gitops.views.templates',
-	DocumentationView = 'gitops.views.documentation',
-}
